Changed output file naming scheme
Changed the output file naming scheme to be based on the RNG seed
instead of the agent output file name passed to -e mode. This works
better because the agent output file name is often pointing to an
entirely different directory, whereas using a constant name based off
of the RNG seed guarantees that the files will be saved in the
shared-population directory with a distinct RNG seed-based name.